agent/requirements.txt
livekit-agents
livekit
livekit-plugins-openai
livekit-plugins-silero
python-dotenv
aiohttp
